Check for the appropriate permissions in File type tests on Windows
Ruby's interface to the permissions on Windows does not map well to
the *nix concept of User, Group, and Other.  On Windows directories
don't get the execute bit, and Ruby cannot manage group, and other via
the standard chmod interfaces.

Because of this, we no longer check that the execute bit is set on
Windows, and use a permission set that will show differences if we
fail to set the permissions on Windows.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>
(cherry picked from commit 447c1171845d8d17e4e684c9508fddecc003d15e)