(#8663) Exclude exec timeout test on Windows
The exec provider is not yet supported for Windows and this test relies
on /bin/sleep. And due to #8410 it attempts to execute the program named
"sleep 1", with variations of "sleep 1".bat, etc. This test needs to be
reworked for Windows when the exec provider is implemented.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>
(cherry picked from commit 0efe900902f0a6528c78c966b73f3eb6c720353b)