maint: Fix inspect help
Running `puppet help inspect` did not result in help:

    Unable to find application 'inspect'.
    err: exit
    err: Try 'puppet help help help' for usage

It turned out that the only reason applications were getting required so
that their help could be found was the LegacyName conversion table in
lib/puppet/util/command_line.rb:7.  Inspect never had a legacy name, so
the help system couldn't find it since it never got required.  Now
instead of checking for the class constant to see if the application has
been loaded, we try to require the application and exit if it's not
found.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>