Do not leak indirector state from apply tests
Since the indirector state persists across tests, we were seeing order
dependent test failures with tests that assumed the default indirector
settings.

Specifically, if the following tests were run in order, the first
would cause failures in the second two:

  spec/unit/application/apply_spec.rb
  spec/unit/node_spec.rb
  spec/integration/node_spec.rb

To protect against this state leakage, we now:

  - reset the Puppet::Node terminus before each test in
    spec/integration/node_spec.rb to ensure we are testing a clean
    environment.

  - reset the Puppet::Node, and Puppet::Node::Facts terminus, and
    cache class after each test in spec/unit/application/apply_spec.rb
    to prevent leakage into other tests.

Since the cache class has the same state leakage problem as the
terminus class, but does not have the same ability to lazily populate
the default when set to nil, we remove the test.  Testing the default
for the cache class would require running the test before all other
tests to ensure there is no state pollution.n