(#9051) Get the compiler out of the ActiveRecord business.
When we started to abstract the system initialization of our database
connections moved out of the parser and AST, into the compiler terminus.

This was excellent progress, and helped abstract away our stored configuration
passing system from the language itself.

Now, as we move StoreConfigs out into a separate and replaceable system we
need to get the compiler out of the business of database initialization
entirely.  That can be fully delegated down to the StoreConfigs terminus and
implementations themselves.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>