(#9051) Dead code elimination in the compiler terminus.
The `update_node_check` method in the compiler terminus used to be used for
compilation checking, but that went the way of the dodo a long time back.

Since nothing and no one in the codebase actually uses this method, and it is
the last lingering bit of database access in the compiler, we can ditch it and
gratefully separate out the compiler from any direct database relationship.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>