(#8140) Add an exec provider for Windows
This provider inherits from the Puppet::Provider::Exec class, and is
very similar to the posix provider in its behavior. This provider
doesn't have the ability to run as a particular user or group and will
fail if that is attempted, but does support setting all other
parameters, as well as autorequires.