(#9051) Implement the `resource` terminus for StoreConfigs.
Add a StoreConfigs terminus to the `resource` indirection, and an ActiveRecord
implementation to back that, using the current StoreConfigs schema.

This is done by extracting the database interaction out of the compiler and
into an appropriate indirection.  Following that, push down the searching into
the terminus itself.

With this we have most of the ability to drop in a replacement for the current
StoreConfigs storage engine.

Along the way, unstub a whole pile of Puppet::Parser::Collector testing so
that it works with the new code and, more importantly, actually tests things
along the way.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>