(#8410) Add a helper to Puppet::Util to determine absoluteness of a path
The method Puppet::Util.absolute_path? accepts a path and an optional
platform (:windows or :posix) as an argument. If no platform is given,
it will default to the local platform. The method will return true if
the path provided is absolute on that platform, and false if it is not.
This method is also being used in Puppet::Util.which in lieu of testing
against a platform-specific regex inline.