(#9051) Make generic tagging imported resource origins.
When StoreConfigs was tied to rails, parser resources were tagged with a
`rails_id` to allow the code to determine if this was the same object we had
previously imported.

Now that we are making that generic, rename that to `collector_id` to make it
clear that it isn't so directly tied to the original purpose.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>