(#8410) Use absolute_path? for Puppet::Parameter::Path validation
A Puppet::Parameter::Path's value is valid if it is fully qualified (or
in the case of an array, if each of its elements is fully qualified).
This class was previously using its own method to validate whether a
path was fully qualified, which was different from the rules applied by
absolute_path?. Now that we have a dedicated method for this,
it is used here for consistency.