(#8667) Write out a list of resources that are managed by puppet agent
Similar to how the Puppet classes are written out each catalog apply,
the list of resources is now being written out to a text file that can
be easily used by tools like MCollective.  This allows tools that do
ad-hoc management of resources to know if they're changing a resource
that puppet manages, and adjust behavior accordingly.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>