(#9051) Port query tests into the indirection.
Now that the parser is using a terminus to search, it shouldn't be in the
business of making assertions about how the Rails level search is constructed
from the input data.

On the other hand, the ActiveRecord resource terminus *should* be in that
business.

This transplants the tests from one to the other, then rewrites the entire
show to make them apply in their new environment.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>