Properly determine file deletion in puppet/unit/util_spec.rb
This test had previously been checking that the path attribute of the
Tempfile used for stdout was nil, in order to determine whether or not
the file had been deleted. However, on Ruby 1.8.5, a Tempfile's data is
not cleared when the file is closed and deleted. Thus, its path wasn't
actually nil and it was failing, even though the file had properly been
deleted.

Now, we just check whether or not the file exists, which will always
directly correspond to whether or not the file has been deleted.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>