(#8412) Add MSI package provider for use with Windows
This provider takes some of its inspiration from the appdmg provider
used with OS X.  It will maintain a list of packages that have been
installed and removed from the system via the provider in a directory
under Puppet's vardir called db/package/msi.  These state files will
be named the same as the resource name with '.yml' appended.  The
state files will be a hash containing the resource name, the install
options used, and the source location of the MSI.

Any properties that a user wishes to provide to the MSI can be
specified as key/value pairs in the install_options parameter.  For
example:

  package { 'mysql':
    provider => msi,
    source => 'E:\mysql.msi',
    ensure => installed,
    install_options => { 'INSTALLDIR' => 'C:\mysql' },
  }

The MSI properties specified by install_options will be appropriately
quoted when invoking msiexec.exe to install the MSI.

Because the source parameter is integral to the functionality of being
able to install and uninstall MSI packages, we also override
validate_source to make sure that the source parameter is always set,
and is not an empty string when using this provider.