(#9174) Provide a helpful error when missing a gem and installing on Windows
We require the sys-admin, win32-process, win32-dir, and win32-service
gems in order to run on Windows, and the win32-dir gem to install.
Previously, failing to require the win32-dir gem would just raise an
exception and fail. Now we catch it, and turn it into a more helpful
explanation of the requirements.