(#8986) Don't su when creating SSH authorized_keys
In order to support use cases where an authorized_key file is written to
a non-standard location, which may not be writable by the user, this patch
removes the step in the flush method that switches users before writing
the authorized_key file to disk. As a result, the authorized_key can now
be written to any location.

This patch does not change the core functionality of the
ssh_authorized_key type.