(#8662) Don't rely on error message to detect UAC capable platform
The call to Win32::Security.elevated_privileges? can raise an
exception when running on a pre-Vista computer or if the process fails
to open its process token.

Previously, we were looking at the exception message to determine
which case it was. However, Windows 2003 and 2003 R2 return different
error codes (and therefore messages) for the pre-Vista case. In 2003,
it returns error code 1 (Incorrect function), but in 2003 R2 it
returns 87 (The parameter is incorrect). Since SUIDManager was only
looking for Incorrect function, SUIDManager.root? would always return
false on 2003 R2.

Ideally, we could just check if the GetTokenInformation Win32 API was
available, and only call it on platforms where it makes sense. But
this API is available on all recent version of Windows. What's new in
Vista and up is the TokenElevation value of the
TOKEN_INFORMATION_CLASS enumeration.

This commit changes the suidmanager to only call GetTokenInformation
when the major kernel version, as reported by facter, is 6.0 or
greater, which corresponds to Vista/2008. See:

http://msdn.microsoft.com/en-us/library/ms724833(v=vs.85).aspx