Fixed #9258 - Puppet Rack logging
Puppet Master as Rack application ignores syslogfacility

Attempting to define a syslogfacility in the config file (under [main]
or [master], makes no difference) or in ARGV in rack/config.ru when
running Puppet as Rack app through Passenger will not apply the setting,
and it will stubbornly log with the default settings to syslog.

This can be traced back to the fact that both daemon.rb and
network/server.rb (during daemonize) call Puppet::Util::Log.reopen, but
when using Rack, network/http/rack.rb will be used, which never does.

This add reopen to rack.rb and a test

Patch thanks to Ronald van Zantvoort.