(#9051) More storeconfigs test cleanup.
We had additional global state modifications out of my testing, where the
cache class and terminus for indirections were changed.  This caused, in some
cases other than "running on my workstation", the tests to fail trying to use
the wrong terminus.

This fixes that, by resetting the state appropriately so we don't leak global
configuration and break each other.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>