(#8413) Only try to catch Process::Error if it's defined
This error class is only defined when using the win32-process gem on
Windows. So rather than always trying to rescue Process::Error (which of
course causes its own error), we build a list of classes to rescue,
which initially only contains Errno::ESRCH, and add Process::Error if
it exists.