Fix tests for #1886 with ActiveRecord 3.x
We didn't properly call the method to test, which lead to a failing
test. However, this couldn't be properly displayed on ruby
installations with an ActiveRecord 3.x installed, as the inspect in
the `after test` part, tried to examine the table, which (obviuosly)
raised a NotConnected exception.

By calling the method the right way, we fix the test and hence also
fix the failing inspection.