(#9051) Move complex collect expression error into terminus.
When the StoreConfig system was extracted from core to a set of termini, most
of the rules about permitted syntax were pushed down into the same place, to
allow them to also be replaced.

One set of restrictions were missed, the limitation that complex search
criteria (like and, or, or parenthetical expressions) were not permitted, and
remained in our parser.

Now, they live in the terminus, and we enforce them only there.  This ensures
that StoreConfigs can be replaced with a back-end that supports complex
collection criteria without other changes to the Puppet core.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>