Fix ActiveRecord handling of symbols in query interpolation.
Most versions of ActiveRecord, including the ones I test with, treat an
interpolated symbol in an SQL query as a simple string, vis:

  ["example = ?", :banana] => "... WHERE example = 'banana' ..."

Some versions, though, preserve the symbol-nature, and instead:

  ["example = ?", :banana] => "... WHERE example = ':banana' ..."
  # ...or some variant of that, where it isn't just the string.

So, to avoid this confusion, manually stringify the symbol name of our
resource type in the query builder.  This gives consistent and expected
behaviour at all times.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>