Remove unnecessary Windows-on-non-Windows-master code for path parameter
This code was taking special care to properly validate and munge the
path parameter in the case where a non-Windows master was compiling a
catalog for a Windows agent. However, validation and munging are only
performed on the agent, so we can simply use the available
platform-specific methods for verifying and splitting the path.