(#2744) Display file diffs through the Puppet log system.
When Puppet generated a diff between the file on disk, it previously just
printed it directly.  This means that the user can view it, but it is lost in
the rest of the system - monitoring, logs, and reports have no visibility of
this.

Better, then, to send it through our regular logging system, so that the
content is visible in all the places that it might be viewed by the user or
monitored by machines.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>