(#9435) Fix absolute path matching for file log destinations
Puppet::Util::Log::DestFile requires that an absolute path be
specified as the log file destination. Previously, it enforced this
requirement using a regex that matched paths starting with '/', which
doesn't work on Windows.

This commit changes the file log destination type to use
Puppet::Util.absolute_path? when determining file absolute-ness.