(#9186) Handle when running under non 'user' contexts
The Sys::Admin.get_user method doesn't work when the username exists
in the WMI Win32_Account class, but not the Win32_UserAccount
class. For example, SYSTEM which is the user context under which the
jenkins agent runs under.

This commit modifies the spec test to query the more general
Win32_Account class. This code will be moved to a more generic
location in a future commit.