(#9186) Fix tests that fail on 2008 when running as SYSTEM
The tests were failing on Jenkins because the SYSTEM account's temp
directory is C:\windows\temp, which on 2008 only has 2 access control
entries (one for Administrators and one for SYSTEM). One test assumed
that newly created file/directories had extra entries, e.g. Users. The
other test assumed that unprotecting the DACL for a newly created
file/directory would cause it to inherit extra access control entries
from its parent. But on 2008, the sids for the parent aces
corresponded to the owner and group, so the sids were not "extra".

I fixed both of these tests to not make assumptions about the access
control entries present for newly created files.